move link types to public link module
