+++ dnp3/src/link/reader.rs
@@ -1,8 +1,9 @@
-use crate::config::{DecodeLevel, LinkErrorMode};
+use crate::config::DecodeLevel;
 use crate::link::display::LinkDisplay;
 use crate::link::error::LinkError;
 use crate::link::header::Header;
 use crate::link::parser::{FramePayload, Parser};
+use crate::link::LinkErrorMode;
 use crate::util::cursor::ReadCursor;
 use crate::util::phys::PhysLayer;
 use std::io::ErrorKind;
