+++ dnp3/src/master/error.rs
@@ -1,8 +1,8 @@
 use crate::app::enums::CommandStatus;
 use crate::app::header::{Iin, Iin2};
 use crate::app::parse::error::ObjectParseError;
-use crate::config::EndpointAddress;
 use crate::link::error::LinkError;
+use crate::link::EndpointAddress;
 use crate::master::association::NoAssociation;
 use crate::master::session::RunError;
 use crate::tokio::sync::mpsc::error::SendError;
