+++ dnp3/src/master/handle.rs
@@ -8,7 +8,7 @@ use crate::app::types::LinkStatusResult;
 use crate::app::types::Timestamp;
 use crate::app::variations::Variation;
 use crate::config::DecodeLevel;
-use crate::config::EndpointAddress;
+use crate::link::EndpointAddress;
 use crate::master::association::{Association, AssociationConfig};
 use crate::master::error::{AssociationError, CommandError, PollError, TaskError, TimeSyncError};
 use crate::master::messages::{AssociationMsg, AssociationMsgType, MasterMsg, Message};
