+++ dnp3/src/master/session.rs
@@ -6,8 +6,8 @@ use crate::app::sequence::Sequence;
 use crate::app::timeout::Timeout;
 use crate::app::types::LinkStatusResult;
 use crate::config::DecodeLevel;
-use crate::config::EndpointAddress;
 use crate::link::error::LinkError;
+use crate::link::EndpointAddress;
 use crate::master::association::{AssociationMap, Next};
 use crate::master::error::TaskError;
 use crate::master::messages::{MasterMsg, Message};
