+++ dnp3/src/master/tasks/restart.rs
@@ -118,7 +118,7 @@ mod tests {
     use crate::app::header::{Control, Iin, ResponseFunction};
     use crate::app::sequence::Sequence;
     use crate::app::variations::{Group52Var1, Group52Var2};
-    use crate::config::EndpointAddress;
+    use crate::link::EndpointAddress;
     use crate::master::association::{Association, AssociationConfig};
     use crate::master::handle::NullHandler;
     use crate::master::tasks::RequestWriter;
