+++ dnp3/src/transport/reader.rs
@@ -1,7 +1,7 @@
 use crate::app::parse::parser::ParsedFragment;
-use crate::config::EndpointAddress;
-use crate::config::{AppDecodeLevel, DecodeLevel, LinkErrorMode};
+use crate::config::{AppDecodeLevel, DecodeLevel};
 use crate::link::error::LinkError;
+use crate::link::{EndpointAddress, LinkErrorMode};
 use crate::outstation::config::Feature;
 use crate::transport::{
     FragmentInfo, LinkLayerMessage, TransportData, TransportRequest, TransportResponse,
