+++ dnp3/src/transport/writer.rs
@@ -1,8 +1,9 @@
 use crate::app::parse::parser::{FragmentDisplay, ParsedFragment};
 use crate::app::EndpointType;
-use crate::config::{DecodeLevel, EndpointAddress};
+use crate::config::DecodeLevel;
 use crate::link::error::LinkError;
 use crate::link::header::AnyAddress;
+use crate::link::EndpointAddress;
 use crate::util::phys::PhysLayer;
 
 /// This type definition is used so that we can mock the transport writer during testing.
